backend/requirements.txt
Django>=3.0,<4.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
mysqlclient==1.4.6
Pillow==6.1.0



